backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
tensorflow==2.18.0
torch==2.5.0
torchaudio==2.5.0
transformers==4.46.1
opencv-python-headless==4.10.0.84
librosa==0.10.2.post1
soundfile==0.12.1
nltk==3.8.1
scikit-learn==1.5.2
numpy==2.1.1
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
# Chart.js via CDN
# Note: Use CPU versions for deployment
